\name{int64-package}
\alias{int64-package}
\docType{package}
\title{
64 bit integer types
}
\description{
64 bit integer types
}
\details{
\tabular{ll}{
Package: \tab int64\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-10-21\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois, Sponsored by Google Open Source Programs Office

Maintainer: Romain Francois <romain@r-enthusiasts.com>
}
\examples{
    as.int64( 1:4 )
    as.int64( c("123456789", "9876543219876543" ) )
}
\keyword{ package }

