\name{sort-methods}
\docType{methods}
\alias{sort-methods}
\alias{sort,ANY-method}
\alias{sort,int64-method}
\alias{sort,uint64-method}
\title{Sorting 64 bits integer vector}
\description{
 Sorting 64 bits integer vector
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
    Standard method (from base)
}

\item{\code{signature(x = "int64")}}{
    Sorting signed 64 bit integer vectors (\code{\linkS4class{int64}}
}

\item{\code{signature(x = "uint64")}}{
    Sorting unsigned 64 bit integer vectors (\code{\linkS4class{uint64}}
}
}}
\keyword{methods}
