% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R, R/ptype.R
\name{vctrs-compat}
\alias{vctrs-compat}
\alias{vec_cast.int64}
\alias{vec_ptype2.int64}
\title{vctrs compatibility functions}
\usage{
\method{vec_cast}{int64}(x, to, ...)

\method{vec_ptype2}{int64}(x, y, ...)
}
\arguments{
\item{x, y}{Objects.}

\item{to}{Type to cast to.}

\item{...}{Used to pass along error message information.}

\item{op}{An arithmetic operator as a string.}
}
\value{
See the corresponding vctrs function for the exact return value.
}
\description{
These functions are the extensions that allow rray objects to
work with vctrs.
}
