% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int64-package.R
\docType{class}
\name{binary-class}
\alias{binary-class}
\alias{show,binary-method}
\title{Class \code{"binary"}}
\description{
Binary representation
}
\section{Objects from the Class}{
 Objects can be created by one of the forms
of the \code{\link{binary}} methods.
}

\examples{

binary( 1:4 )

}
\author{
Romain Francois, sponsored by the Google Open Source Programs Office
}
\keyword{classes}
