% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int64-package.R
\docType{package}
\name{int64-package}
\alias{int64-package}
\title{64 bit integer types}
\description{
An implementation of 64-bit integer vectors for R.
}
\examples{

    as.int64( 1:4 )
    as.int64( c("123456789", "9876543219876543" ) )

}
\author{
Romain Francois, Sponsored by Google Open Source Programs Office, Balasubramanian Narasimhan
Maintainer: Balasubramanian Narasimhan \href{mailto:naras@stanford.edu}{naras@stanford.edu}
}
\keyword{package}
