% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int64-package.R
\docType{package}
\name{int64-package}
\alias{int64}
\alias{int64-package}
\title{int64: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
