% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int64-package.R
\docType{package}
\name{int64-package}
\alias{int64}
\alias{int64-package}
\title{int64: Large Integer Types}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
