% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int64.R
\name{int64}
\alias{int64}
\title{Creates new int64 vectors of a given length}
\usage{
int64(length = 0L)
}
\arguments{
\item{length}{Length of the resulting vector}
}
\value{
A new \code{\linkS4class{int64}} vector of the given length
}
\description{
Creates new \code{\linkS4class{int64}} vectors of a given length
}
\examples{

x <- int64(10L)

}
\seealso{
\code{\link{as.int64}} for converting integer or character vectors
into \code{\linkS4class{int64}} vectors.

\code{\link{uint64}} for unsigned 64 bit integer vectors, i.e. of class
\code{\linkS4class{uint64}}
}
\author{
Romain Francois, Sponsored by the Google Open Source Programs Office
}
\keyword{manip}
