% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int64-package.R
\docType{methods}
\name{sort-methods}
\alias{sort-methods}
\alias{sort,ANY-method}
\alias{sort,int64-method}
\alias{sort,uint64-method}
\title{Sorting 64 bits integer vector}
\description{
Sorting 64 bits integer vector
}
\section{Methods}{
 \describe{

\item{list("signature(x = \"ANY\")")}{ Standard method (from base) }

\item{list("signature(x = \"int64\")")}{ Sorting signed 64 bit integer
vectors (\code{\linkS4class{int64}} }

\item{list("signature(x = \"uint64\")")}{ Sorting unsigned 64 bit integer
vectors (\code{\linkS4class{uint64}} } }
}

\keyword{methods}
